refactores code
